fixing bugs
so firstly, removed an invisible line at the top. then made it so a user can enter 0, 1, and ?. earlier, you could add nubmers like 8 and letters llike c and it did give results. i've known this for a while but i knew someone would probably say, @xVoid879 there's a bug etc. so i just fixed it. made it so it also allows whtiespace (even though you shouldn't) just in case it doesn't work if a user creates a new then runs the pattern.

if (!/^[01?\s]+$/.test(patternText)) {
  resultsDiv.textContent = 'Please only use 0, 1, and ?';
  return;
}